feat: 나의 투표내역 조회 API 구현

+++ src/main/java/motgolla/domain/vote/dto/response/VoteDetailResponse.java
@@ -22,7 +22,7 @@ public class VoteDetailResponse {
         private Long recordId;
         private int voteCount;
         private int totalVotes;
-        private Double percentage;
+        private Integer percentage;
         private String imageUrl;
     }
 }
